Add infrastructure to generate a ToC
A task is provided to use the "markdown-toc" tool to generate a table of contents for Markdown files in the project.

A GitHub Actions workflow is included to automatically run the task. The workflow is triggered on any push or pull that
changes relevant project files, in order to avoid the introduction of problems with the project filesystem. It is also
triggered periodically, in order to catch breakage caused by external changes.