Add infrastructure to check spelling
Tasks are provided to check for commonly misspelled words in the project files, and correct those misspellings.

A GitHub Actions workflow is included to automatically run the spell check task. The workflow is triggered on any push
or pull that changes relevant project files, in order to avoid the introduction of problems with the project filesystem.
It is also triggered periodically, in order to catch breakage caused by external changes.