Add infrastructure to check for problems with npm configuration files
Tasks are provided to check for problems in the project's npm configuration files:

* Validate package.json against its JSON schema.
* Check for forgotten package-lock.json syncs.

A GitHub Actions workflow is included to automatically run the tasks. The workflow is triggered on any push or pull that
changes relevant project files, in order to avoid the introduction of problems with the project filesystem. It is also
triggered periodically, in order to catch breakage caused by external changes.